Added new rpackage tasks: uploadCRAN, emailCRAN, and publishCRAN, which simply depends on the two others.  These allow users to simply and easily publish packages on CRAN with gradle.
Also added the RExec task, which can run an R script from a file,
evaluate an R expression, or run an `R CMD` command.  In any of the
cases, `args` can be used to pass additional arguments to R, e.g.
`--vanilla` for running R code or `--as-cran` for `R CMD check`.  All of
the tasks in `rpackage` were updated to use RExec, so you can easily add
options.